Refactored code : added readme file, reorganised project structure into publisher and consumer, moved dtype management to sql insertion script, cleaned unused code
