+++ data_publisher/main.py
@@ -1,0 +1,7 @@
+"""Script to write local csv files from simulations."""
+from publisher import emit_truck_info
+
+
+if __name__ == '__main__':
+    while True:
+        emit_truck_info()
