+++ data_publisher/publisher.py
@@ -2,12 +2,8 @@ import pika
 import json
 import time
 from truck import Truck
-from random import randint
 from datetime import datetime
 
-def stringify_position(position):
-    template = 'POINT(%f %f)'
-    return template % (position[0], position[1])
 
 def emit_message(payload):
     connection = pika.BlockingConnection(
@@ -21,16 +17,13 @@ def emit_message(payload):
                           body=json.dumps(payload),
                           )
     print(" [x] Message publisher %r" % payload)
-    print('trying close')
     connection.close()
-    print('closed')
 
 
 def emit_truck_info():
-    print('new emit')
-    id_truck = randint(0, 100)
-    id_driver = randint(0, 100)
-    id_itinerary = randint(0, 100)
+    id_truck = 1
+    id_driver = 2
+    id_itinerary = 2
     truck = Truck(id_truck)
     departure, arrival = truck.drive()
     journey_ended = False
@@ -42,15 +35,10 @@ def emit_truck_info():
                     "idtruck": id_truck,
                     "status": 0,
                     "iditinerary": id_itinerary,
-                    "position": stringify_position(position),
+                    "position": position,
                     "timestamp": datetime.now().strftime("%m/%d/%Y %H:%M:%S")
                   }
         journey_ended = position == arrival
         emit_message(payload)
-        print('emited')
         t += 1
         time.sleep(1)
-
-
-while True:
-    emit_truck_info()
