+++ data_publisher/truck.py
