--- infos/messages_structure.json
@@ -1,15 +1,0 @@
-{"position":
-    {"id_driver":"1",
-     "id_truck": "1",
-     "status": "0",
-     "id_itinerary":"1",
-     "position": "coord"}
-}
-
-
-{"itinerary": {
-    "id":"1",
-    "mission":"text",
-    "departure":"coord",
-    "arrival":"coord"
-}}
