+++ readme.md
@@ -1,0 +1,43 @@
+## Truck Simulator 2020
+
+Starting a readme
+
+Go jap !
+
+Table of contents
+
+- [Truck-simulator](#dev2019gojap)
+  - [Client](#client)
+    - [Overview](#overview)
+    - [Used Libraries :](#used-libraries)
+    - [Installation :](#installation)
+    
+## Project Overview
+
+### Installation :
+
+- `docker` 
+- `pipenv` 
+
+## Data publisher
+
+### Overview
+
+The data generator simulates the position of the trucks
+
+### Used Libraries :
+
+- [Pika](https://pika.com/) (RabbitMQ)
+
+## Data consumer
+
+### Overview
+
+Blablabla
+
+### Used Libraries :
+
+- [Pika](https://pika.com/) (RabbitMQ)
+
+### Run database
+- `docker-compose up`
